move everything to scanner
